refactor, fix build
