+++ packages/relative-values/src/values/types.ts
@@ -8,3 +8,13 @@ export type RelativeValue = Readonly<{
 }>;
 
 export type RelativeValueResult = result<RelativeValue, string>;
+
+export type ModelCache = {
+  id: string;
+  relativeValues: Record<string, Record<string, RelativeValueResult>>;
+};
+
+export type CatalogCache = {
+  id: string;
+  models: ModelCache[];
+};
